Changed texture format from RGBA32F to RGBA8

+++ OpenGLRendering/OGLTexture.cpp
@@ -46,7 +46,7 @@ TextureBase* OGLTexture::RGBATextureFromData(char* data, int width, int height, 
 
 	glBindTexture(GL_TEXTURE_2D, tex->texID);
 
-	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, width, height, 0, sourceType, GL_UNSIGNED_BYTE, data);
+	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA8, width, height, 0, sourceType, GL_UNSIGNED_BYTE, data);
 
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
